fw-5870, rename mtd tasks file
